# server/nodejs/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.4"
  title: "Syscoin API"
host: "localhost:8001"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /escrowacknowledge:
    post:
      tags:
      - "Escrow"
      description: "Acknowledge escrow payment as seller of offer. Deducts qty of\
        \ offer and increases number of sold inventory."
      operationId: "escrowacknowledge"
      parameters:
      - name: "escrowguid"
        in: "query"
        required: true
        type: "string"
      - name: "witness"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowcreaterawtransaction:
    post:
      tags:
      - "Escrow"
      description: "Creates raw transaction for escrow refund or release, sign the\
        \ output raw transaction and pass it via the rawtx parameter to escrowrelease.\
        \ Type is 'refund' or 'release'. Third parameter is array of input (txid,\
        \ vout, amount) pairs to be used to fund escrow payment. User role represents\
        \ either 'seller', 'buyer' or 'arbiter', represents who signed for the payment\
        \ of the escrow. 'seller' or 'arbiter' is valid for type 'refund', while 'buyer'\
        \ or 'arbiter' is valid for type 'release'. You only need to provide this\
        \ parameter when calling escrowrelease or escrowrefund."
      operationId: "escrowcreaterawtransaction"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowCreateRawTransactionRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowcompleterefund:
    post:
      tags:
      - "Escrow"
      description: "Completes an escrow refund by creating the escrow complete refund\
        \ transaction on syscoin blockchain."
      operationId: "escrowcompleterefund"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowCompleteRefundRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowcompleterelease:
    post:
      tags:
      - "Escrow"
      description: "Completes an escrow release by creating the escrow complete release\
        \ transaction on syscoin blockchain."
      operationId: "escrowcompleterelease"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowCompleteReleaseRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowfeedback:
    post:
      tags:
      - "Escrow"
      description: "Send feedback for primary and secondary users in escrow, depending\
        \ on who you are. Ratings are numbers from 1 to 5. User From and User To is\
        \ either 'buyer', 'seller', 'reseller', or 'arbiter'."
      operationId: "escrowfeedback"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowFeedbackRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowfilter:
    get:
      tags:
      - "Escrow"
      description: "scan and filter escrows"
      operationId: "escrowfilter"
      parameters:
      - name: "query"
        in: "query"
        description: "Generic filter query to pass into syscoinquery"
        required: true
        type: "string"
      - name: "count"
        in: "query"
        description: "The number of results to return"
        required: false
        type: "string"
      - name: "from"
        in: "query"
        description: "Show results from this GUID [from], empty means first"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EscrowIndex"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowbidhistory:
    get:
      tags:
      - "Escrow"
      description: "scan and filter bids on escrow contracts"
      operationId: "escrowbidhistory"
      parameters:
      - name: "query"
        in: "query"
        description: "Generic filter query to pass into syscoinquery"
        required: true
        type: "string"
      - name: "count"
        in: "query"
        description: "The number of results to return"
        required: false
        type: "string"
      - name: "from"
        in: "query"
        description: "Show results from this GUID [from], empty means first"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EscrowBidIndex"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowfeedbackhistory:
    get:
      tags:
      - "Escrow"
      description: "scan and filter feedbacks and ratings"
      operationId: "escrowfeedbackhistory"
      parameters:
      - name: "query"
        in: "query"
        description: "Generic filter query to pass into syscoinquery"
        required: true
        type: "string"
      - name: "count"
        in: "query"
        description: "The number of results to return"
        required: false
        type: "string"
      - name: "from"
        in: "query"
        description: "Show results from this GUID [from], empty means first"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EscrowFeedbackIndex"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowinfo:
    get:
      tags:
      - "Escrow"
      description: "Show stored values of a single escrow"
      operationId: "escrowinfo"
      parameters:
      - name: "escrow"
        in: "query"
        description: "GUID of escrow"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Escrow"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowbid:
    post:
      tags:
      - "Escrow"
      description: "Bid on an auction."
      operationId: "escrowbid"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowBidRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
...
  /escrownew:
    post:
      tags:
      - "Escrow"
      description: "Create new arbitrated Syscoin escrow."
      operationId: "escrownew"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowNewRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowrefund:
    post:
      tags:
      - "Escrow"
      description: "Refunds escrow funds back to buyer, buyer needs to sign the output\
        \ transaction and send to the network. User role represents either 'seller'\
        \ or 'arbiter'. Enter in rawTx if this is an external payment refund."
      operationId: "escrowrefund"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowRefundRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
  /escrowrelease:
    post:
      tags:
      - "Escrow"
      description: "Releases escrow funds to seller, seller needs to sign the output\
        \ transaction and send to the network. User role represents either 'buyer'\
        \ or 'arbiter'. Enter in rawTx if this is an external payment release."
      operationId: "escrowrelease"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EscrowReleaseRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Escrow"
    x-swagger-router-controller: "rpc"
...
securityDefinitions:
  token:
    type: "apiKey"
    name: "token"
    in: "header"
definitions:
...
  ErrorResponse:
    required:
    - "message"
    properties:
      message:
        type: "string"
...
  Escrow:
    properties:
      _id:
        type: "string"
      time:
        type: "number"
      seller:
        type: "string"
      arbiter:
        type: "string"
      buyer:
        type: "string"
      witness:
        type: "string"
      offer:
        type: "string"
      offer_price:
        type: "string"
      reseller:
        type: "string"
      quantity:
        type: "number"
      total_with_fee:
        type: "number"
      total_without_fee:
        type: "number"
      bid_in_offer_currency_per_unit:
        type: "number"
      total_or_bid_in_payment_option_per_unit:
        type: "number"
      buynow:
        type: "boolean"
      commission:
        type: "number"
      arbiterfee:
        type: "number"
      networkfee:
        type: "number"
      witnessfee:
        type: "number"
      shipping:
        type: "number"
      deposit:
        type: "number"
      currency:
        type: "string"
      exttxid:
        type: "string"
      escrowaddress:
        type: "string"
      paymentoption:
        type: "string"
      redeem_txid:
        type: "string"
      redeem_script:
        type: "string"
      txid:
        type: "string"
      height:
        type: "number"
      role:
        type: "string"
      expired:
        type: "boolean"
      status:
        type: "string"
    example:
      seller: "aeiou"
      total_without_fee: 5.962133916683182377482808078639209270477294921875
      role: "aeiou"
      redeem_txid: "aeiou"
      offer: "aeiou"
      expired: true
      shipping: 4.1456080298839363962315474054776132106781005859375
      arbiterfee: 9.301444243932575517419536481611430644989013671875
      commission: 7.061401241503109105224211816675961017608642578125
      currency: "aeiou"
      escrowaddress: "aeiou"
      height: 1.231513536777255612975068288506008684635162353515625
      total_with_fee: 1.46581298050294517310021547018550336360931396484375
      quantity: 6.02745618307040320615897144307382404804229736328125
      witnessfee: 2.027123023002321833274663731572218239307403564453125
      total_or_bid_in_payment_option_per_unit: 2.3021358869347654518833223846741020679473876953125
      paymentoption: "aeiou"
      networkfee: 3.61607674925191080461672754609026014804840087890625
      reseller: "aeiou"
      txid: "aeiou"
      buyer: "aeiou"
      offer_price: "aeiou"
      witness: "aeiou"
      arbiter: "aeiou"
      buynow: true
      exttxid: "aeiou"
      deposit: 7.3862819483858839220147274318151175975799560546875
      redeem_script: "aeiou"
      _id: "aeiou"
      time: 0.80082819046101150206595775671303272247314453125
      bid_in_offer_currency_per_unit: 5.63737665663332876420099637471139430999755859375
      status: "aeiou"
  EscrowIndex:
    properties:
      _id:
        type: "string"
      offer:
        type: "string"
      escrow:
        type: "string"
      height:
        type: "number"
      seller:
        type: "string"
      arbiter:
        type: "string"
      buyer:
        type: "string"
    example:
      offer: "aeiou"
      seller: "aeiou"
      arbiter: "aeiou"
      escrow: "aeiou"
      _id: "aeiou"
      height: 0.80082819046101150206595775671303272247314453125
      buyer: "aeiou"
  EscrowBidIndex:
    properties:
      _id:
        type: "string"
      offer:
        type: "string"
      escrow:
        type: "string"
      height:
        type: "number"
      bidder:
        type: "string"
      bid_in_offer_currency_per_unit:
        type: "string"
      bid_in_payment_option_per_unit:
        type: "string"
      witness:
        type: "string"
      status:
        type: "string"
    example:
      offer: "aeiou"
      bid_in_payment_option_per_unit: "aeiou"
      witness: "aeiou"
      bidder: "aeiou"
      escrow: "aeiou"
      _id: "aeiou"
      bid_in_offer_currency_per_unit: "aeiou"
      height: 0.80082819046101150206595775671303272247314453125
      status: "aeiou"
  EscrowFeedbackIndex:
    properties:
      _id:
        type: "string"
      offer:
        type: "string"
      escrow:
        type: "string"
      txid:
        type: "string"
      time:
        type: "number"
      rating:
        type: "number"
      feedbackuserfrom:
        type: "string"
      feedbackuserto:
        type: "string"
      feedback:
        type: "string"
    example:
      offer: "aeiou"
      feedback: "aeiou"
      feedbackuserfrom: "aeiou"
      rating: 6.02745618307040320615897144307382404804229736328125
      escrow: "aeiou"
      txid: "aeiou"
      feedbackuserto: "aeiou"
      _id: "aeiou"
      time: 0.80082819046101150206595775671303272247314453125
  EscrowRefundRequest:
    required:
    - "escrowguid"
    - "rawtx"
    - "userrole"
    properties:
      escrowguid:
        type: "string"
      userrole:
        type: "string"
      rawtx:
        type: "string"
      witness:
        type: "string"
  EscrowReleaseRequest:
    required:
    - "escrowguid"
    - "rawtx"
    - "userrole"
    properties:
      escrowguid:
        type: "string"
      userrole:
        type: "string"
      rawtx:
        type: "string"
      witness:
        type: "string"
...
  EscrowCreateRawTransactionRequest:
    required:
    - "escrowguid"
    - "inputs"
    - "type"
    properties:
      type:
        type: "string"
      escrowguid:
        type: "string"
      inputs:
        type: "array"
        items:
          $ref: "#/definitions/GetAddressUTXOsEntry"
      role:
        type: "string"
  EscrowCompleteRefundRequest:
    required:
    - "escrowguid"
    - "rawtx"
    properties:
      escrowguid:
        type: "string"
      rawtx:
        type: "string"
      witness:
        type: "string"
  EscrowCompleteReleaseRequest:
    required:
    - "escrowguid"
    - "rawtx"
    properties:
      escrowguid:
        type: "string"
      rawtx:
        type: "string"
      witness:
        type: "string"
  EscrowBidRequest:
    required:
    - "alias"
    - "bid_in_offer_currency"
    - "bid_in_payment_option"
    - "escrow"
    properties:
      alias:
        type: "string"
        description: "An alias you own."
      escrow:
        type: "string"
        description: "Escrow GUID to place bid on."
      bid_in_payment_option:
        type: "string"
        description: "Amount to bid on offer through escrow. Bid is in payment option\
          \ currency. Example, If offer is paid in SYS and you have deposited 10 SYS\
          \ in escrow and would like to increase your total bid to 14 SYS enter 14\
          \ here. It is per unit of purchase."
      bid_in_offer_currency:
        type: "string"
        description: "Converted value of bid_in_payment_option from paymentOption\
          \ currency to offer currency. For example, offer is priced in USD and purchased\
          \ in BTC, this field will be the BTC/USD value. It is per unit of purchase."
      witness:
        type: "string"
        description: "Witness alias name that will sign for web-of-trust notarization\
          \ of this transaction."
  EscrowNewRequest:
    required:
    - "alias"
    - "arbiter"
    - "buynow"
    - "getamountandaddress"
    - "offer"
    - "quantity"
    - "total_in_payment_option"
    properties:
      getamountandaddress:
        type: "string"
        description: "True or False. Get deposit and total escrow amount aswell as\
          \ escrow address for funding. If buynow is false pass bid amount in bid_in_payment_option\
          \ to get total needed to complete escrow. If buynow is true amount is calculated\
          \ based on offer price and quantity."
      alias:
        type: "string"
        description: "An alias you own."
      arbiter:
        type: "string"
        description: "Alias of arbiter."
      offer:
        type: "string"
        description: "GUID of offer that this escrow is managing."
      quantity:
        type: "string"
        description: "Quantity of items to buy of offer."
      buynow:
        type: "string"
        description: "Specify whether the escrow involves purchasing offer for the\
          \ full offer price if set to true, or through a bidding auction if set to\
          \ false. If buynow is false, an initial deposit may be used to secure a\
          \ bid if required by the seller."
      total_in_payment_option:
        type: "string"
        description: "Total amount of the offer price. Amount is in paymentOption\
          \ currency. It is per unit of purchase."
      shipping:
        type: "string"
        description: "Amount to add to shipping for merchant. Amount is in paymentOption\
          \ currency. Example, If merchant requests 0.1 BTC for shipping and escrow\
          \ is paid in BTC, enter 0.1 here. Default is 0. Buyer can also add shipping\
          \ using escrowaddshipping upon merchant request."
      networkfee:
        type: "string"
        description: "Network fee in satoshi per byte for the transaction. Generally\
          \ the escrow transaction is about 400 bytes. Default is 25 for SYS or ZEC\
          \ and 250 for BTC payments."
      arbiterfee:
        type: "string"
        description: "Arbiter fee in fractional amount of the amount_in_payment_option\
          \ value. For example 0.75% is 0.0075 and represents 0.0075*amount_in_payment_option\
          \ satoshis paid to arbiter in the event arbiter is used to resolve a dispute.\
          \ Default and minimum is 0.005."
      witnessfee:
        type: "string"
        description: "Witness fee in fractional amount of the amount_in_payment_option\
          \ value. For example 0.3% is 0.003 and represents 0.003*amount_in_payment_option\
          \ satoshis paid to witness in the event witness signs off on an escrow through\
          \ any of the following calls escrownew/escrowbid/escrowrelease/escrowrefund.\
          \ Default is 0."
      exttx:
        type: "string"
        description: "External transaction ID if paid with another blockchain."
      paymentoption:
        type: "string"
        description: "If extTx is defined, specify a valid payment option used to\
          \ make payment. Default is SYS."
      bid_in_payment_option:
        type: "string"
        description: "Initial bid amount you are willing to pay escrow for this offer.\
          \ Amount is in paymentOption currency. It is per unit of purchase. If buynow\
          \ is set to true, this value is disregarded."
      bid_in_offer_currency:
        type: "string"
        description: "Converted value of bid_in_payment_option from paymentOption\
          \ currency to offer currency. For example, offer is priced in USD and purchased\
          \ in BTC, this field will be the BTC/USD value. If buynow is set to true,\
          \ this value is disregarded."
      witness:
        type: "string"
        description: "Witness alias name that will sign for web-of-trust notarization\
          \ of this transaction."
...
  EscrowFeedbackRequest:
    required:
    - "escrowguid"
    - "feedback"
    - "rating"
    - "userfrom"
    - "userto"
    properties:
      escrowguid:
        type: "string"
      userfrom:
        type: "string"
      feedback:
        type: "string"
      rating:
        type: "string"
      userto:
        type: "string"
      witness:
        type: "string"
...
  GetAddressUTXOsEntry:
    properties:
      address:
        type: "string"
      txid:
        type: "string"
      outputIndex:
        type: "number"
      script:
        type: "string"
      satoshis:
        type: "number"
      height:
        type: "number"
    example:
      outputIndex: 0.80082819046101150206595775671303272247314453125
      address: "aeiou"
      txid: "aeiou"
      script: "aeiou"
      satoshis: 6.02745618307040320615897144307382404804229736328125
      height: 1.46581298050294517310021547018550336360931396484375